+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Allow setting ASG max scaling to 0 if ASG min is also set to 0.
+
 ## [3.6.0] - 2021-11-26
 
 ### Added